+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v2.1.0 (2021-12-04)
 
 - Other bug fixes with language rules.
+- Code refactoring
 - Bug fix with English language 0 seconds ago.
 
 ```diff